requirements.txt
neat-python==0.92
pygame==2.6.1
numpy==1.26.4
